+++ packages/orq-rc/src/orq_poc_python_multi_env_version/sdk.py
@@ -10,10 +10,11 @@ from orq_poc_python_multi_env_version import models, utils
 from orq_poc_python_multi_env_version._hooks import SDKHooks
 from orq_poc_python_multi_env_version.contacts import Contacts
 from orq_poc_python_multi_env_version.deployments_sdk import DeploymentsSDK
-from orq_poc_python_multi_env_version.evals import Evals
 from orq_poc_python_multi_env_version.evaluators import Evaluators
 from orq_poc_python_multi_env_version.feedback import Feedback
 from orq_poc_python_multi_env_version.files import Files
+from orq_poc_python_multi_env_version.prompt import Prompt
+from orq_poc_python_multi_env_version.prompts import Prompts
 from orq_poc_python_multi_env_version.remoteconfig import Remoteconfig
 from orq_poc_python_multi_env_version.types import OptionalNullable, UNSET
 from typing import Any, Callable, Dict, Optional, Union
@@ -29,7 +30,8 @@ class Orq(BaseSDK):
     deployments: DeploymentsSDK
     files: Files
     evaluators: Evaluators
-    evals: Evals
+    prompt: Prompt
+    prompts: Prompts
     remoteconfig: Remoteconfig
 
     def __init__(
@@ -116,7 +118,8 @@ class Orq(BaseSDK):
         self.deployments = DeploymentsSDK(self.sdk_configuration)
         self.files = Files(self.sdk_configuration)
         self.evaluators = Evaluators(self.sdk_configuration)
-        self.evals = Evals(self.sdk_configuration)
+        self.prompt = Prompt(self.sdk_configuration)
+        self.prompts = Prompts(self.sdk_configuration)
         self.remoteconfig = Remoteconfig(self.sdk_configuration)
 
     def __enter__(self):